requirements.txt
amqp==5.3.1
anyio==4.8.0
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
dj-config-url==0.1.1
dj-database-url==2.3.0
Django==4.2.14
django-celery-beat==2.7.0
django-js-asset==2.2.0
django-timezone-field==7.1
django-tinymce==4.1.0
exceptiongroup==1.2.2
googletrans==4.0.0rc1
gunicorn==23.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
kombu==5.4.2
packaging==24.2
pillow==11.1.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
python-crontab==3.2.0
python-dateutil==2.9.0.post0
redis==5.2.1
requests==2.32.3
rfc3986==1.5.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.8.2
